added search options

+++ migrations/01_schema.sql
@@ -52,7 +52,7 @@ CREATE TABLE property_reviews (
   guest_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
   property_id INTEGER REFERENCES properties(id) ON DELETE CASCADE,
   reservation_id INTEGER REFERENCES reservations(id) ON DELETE CASCADE,
-  rating INTEGER,
+  rating REAL,
   message VARCHAR(225)
 );
